Refactor output parameters
Remove `--output` parameter, add `--text` parameter for text output into a file, and allow all of them, `--html`, `--markdown` and `--text`, at the same time instead of only one.

The `--output` parameter is not needed because the same functionality is provided by `--html`, `--markdown`, and `--text` in a more intuitive way.