Add support for outputting to Asciidoc format
Asciidoc is very similar to Markdown, but it has a few differences:

- It uses `=` for headers instead of `#`
- It uses multiple `*` for nested lists instead of indents
- It uses `\n'''\n` for horizontal rule instead of `---\n`
- Blocks quotes in Asciidoc are done like this:
  ```
  ----
  QUOTED TEXT
  ACROSS MULTIPLE LINES
  ----
  ```
  versus how it done in Markdown
  ```
  > QUOTED TEXT
  > ACROSS MULTIPLE LINES
  ```

Due to these similarities, they basically share the exact same render code!
The core implementation of MarkdownRender was moved to an abstract base
class MarkdownRenderBase (with abstract methods for getting the symbols for
LI, HR, etc.), with MarkdownRender now just defining those abstract
methods. AsciidocRender also extends the same base class, and also
overrides the blockquote method since blockquoting is done quite
differently and is not a simple symbol replacement.